// src/app/modules/staff/staff-information-system/model/staff-leave-request-view-model.ts
/**
 * Teacher API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StaffLeaveRequestViewModel
 */
export interface StaffLeaveRequestViewModel {
    /**
     * StaffId
     */
    staffId?: number;
    /**
     * Reason
     */
    reason?: string;
    /**
     * FromDate
     */
    fromDate?: Date;
    /**
     * ToDate
     */
    toDate?: Date;
    /**
     * Gets or Sets FileNamewithext
     */
    fileNamewithext?: string;
    /**
     * Gets or Sets FileValue
     */
    fileValue?: string;
}